CV-Tools/requirements.txt
opencv-python~=4.8.1.78
numpy~=1.26.2
moviepy~=1.0.3

Pillow~=10.0.1